+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SeatSelectionActivity"
+            android:name=".ui.seatselectionactivity.SeatSelectionActivity"
             android:exported="false" />
         <activity
             android:name=".ui.MovieBookingCheckActivity"
