+++ app/src/main/res/layout/activity_seat_selection.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/seat_background"
-    tools:context=".ui.SeatSelectionActivity">
+    tools:context=".ui.seatselectionactivity.SeatSelectionActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/lo_seat_table"
